+++ pkg/utils/kubernetes/resources/deployments.go
@@ -2,6 +2,8 @@ package resources
 
 import (
 	"fmt"
+	"sort"
+	"strings"
 	"sync"
 
 	"github.com/samber/lo"
@@ -58,6 +60,12 @@ type GenerateNewDeploymentForControlPlaneParams struct {
 	ServiceAccountName             string
 	AdminMTLSCertSecretName        string
 	AdmissionWebhookCertSecretName string
+	// WatchNamespaces contains the namespaces to watch for resources.
+	// If not nil, the controller will only watch for resources in the specified namespaces.
+	// This list has been verified (and possibly filtered down) by inspecting
+	// the WatchNamespaces from the spec for ReferenceGrants in these namespaces.
+	// If a namespace did not have a ReferenceGrant, it will not be in this list.
+	WatchNamespaces []string
 }
 
 // GenerateNewDeploymentForControlPlane generates a new Deployment for the ControlPlane
@@ -120,6 +128,7 @@ func GenerateNewDeploymentForControlPlane(params GenerateNewDeploymentForControl
 		}
 		deployment.Spec.Template = *patchedPodTemplateSpec
 	}
+	setWatchNamespace(&deployment.Spec.Template.Spec.Containers[0], params.WatchNamespaces)
 
 	k8sutils.SetOwnerForObject(deployment, params.ControlPlane)
 
@@ -132,6 +141,22 @@ func GenerateNewDeploymentForControlPlane(params GenerateNewDeploymentForControl
 	}
 
 	return deployment, nil
+}
+
+func setWatchNamespace(
+	c *corev1.Container,
+	watchNamespaces []string,
+) {
+	if len(watchNamespaces) == 0 {
+		return
+	}
+
+	const watchNamespaceEnvVarName = "CONTROLLER_WATCH_NAMESPACE"
+	sort.Strings(watchNamespaces)
+	c.Env = append(c.Env, corev1.EnvVar{
+		Name:  watchNamespaceEnvVarName,
+		Value: strings.Join(watchNamespaces, ","),
+	})
 }
 
 // GenerateContainerForControlPlaneParams is a parameter struct for GenerateControlPlaneContainer function.
@@ -181,6 +206,7 @@ func GenerateControlPlaneContainer(params GenerateContainerForControlPlaneParams
 			Protocol:      corev1.ProtocolTCP,
 		})
 	}
+
 	return c
 }